requirements.txt
flask
MarkupSafe==2.1.3
charset-normalizer==3.2.0
colorama==0.4.6
cycler==0.11.0
filelock==3.12.2
fonttools==4.42.0
idna==3.4
jinja2==3.1.2
kiwisolver==1.4.4
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
packaging==23.1
piexif==1.1.3
py-cpuinfo==9.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
scipy==1.11.1
six==1.16.0
sympy==1.12
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.6
blinker==1.6.2
click==8.1.6
flask==2.3.2
itsdangerous==2.1.2
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
frozenlist==1.4.0
multidict==6.0.4
openai==0.27.8
yarl==1.9.2
GitPython>=3.1.24
ensemble_boxes
super-gradients
requests
opencv-python
pillow
torchvision
ultralytics
piexif
pymongo
passlib
flask_cors
waitress
twilio
openai